+++ packages/cli/src/commands/index.ts
@@ -138,7 +138,6 @@ export function registerCommands(program: Command) {
       (opt, arr: string[]) => [...arr, opt],
       [],
     )
-    .option('--scope <scope>', 'The scope to use for new packages')
     .option(
       '--npm-registry <URL>',
       'The package registry to use for new packages',
@@ -148,6 +147,10 @@ export function registerCommands(program: Command) {
       'The version to use for any new packages (default: 0.1.0)',
     )
     .option('--no-private', 'Do not mark new packages as private')
+    .option(
+      '--do-not-edit-packages',
+      'Do not edit packages/app and packages/backend',
+    )
     .action(lazy(() => import('./new/new').then(m => m.default)));
 
   registerScriptCommand(program);