+++ packages/cli/src/lib/new/factories/frontendPlugin.test.ts
@@ -89,14 +89,18 @@ describe('frontendPlugin factory', () => {
 
     expect(output).toEqual([
       '',
-      'Creating frontend plugin backstage-plugin-test',
+      'Creating frontend plugin @janus-idp/backstage-plugin-test',
       'Checking Prerequisites:',
       `availability  plugins${sep}test`,
       'creating      temp dir',
       'Executing Template:',
       'copying       .eslintrc.js',
       'templating    README.md.hbs',
+      'templating    app-config.janus-idp.yaml.hbs',
+      'copying       config.d.ts',
       'templating    package.json.hbs',
+      'templating    tsconfig.json.hbs',
+      'templating    turbo.json.hbs',
       'templating    index.tsx.hbs',
       'templating    index.ts.hbs',
       'templating    plugin.test.ts.hbs',
@@ -119,14 +123,14 @@ describe('frontendPlugin factory', () => {
       fs.readJson('/root/packages/app/package.json'),
     ).resolves.toEqual({
       dependencies: {
-        'backstage-plugin-test': '^1.0.0',
+        '@janus-idp/backstage-plugin-test': '^1.0.0',
       },
     });
 
     await expect(fs.readFile('/root/packages/app/src/App.tsx', 'utf8')).resolves
       .toBe(`
 import { createApp } from '@backstage/app-defaults';
-import { TestPage } from 'backstage-plugin-test';
+import { TestPage } from '@janus-idp/backstage-plugin-test';
 
 const router = (
   <FlatRoutes>
@@ -176,7 +180,6 @@ const router = (
     jest.spyOn(Task, 'forCommand').mockResolvedValue();
 
     await frontendPlugin.create(options, {
-      scope: 'internal',
       private: true,
       isMonoRepo: true,
       defaultVersion: '1.0.0',
@@ -188,14 +191,14 @@ const router = (
       fs.readJson('/root/packages/app/package.json'),
     ).resolves.toEqual({
       dependencies: {
-        '@internal/plugin-test': '^1.0.0',
+        '@janus-idp/backstage-plugin-test': '^1.0.0',
       },
     });
 
     await expect(fs.readFile('/root/packages/app/src/App.tsx', 'utf8')).resolves
       .toBe(`
 import { createApp } from '@backstage/app-defaults';
-import { TestPage } from '@internal/plugin-test';
+import { TestPage } from '@janus-idp/backstage-plugin-test';
 
 const router = (
   <FlatRoutes>