+++ plugins/notifications-backend/src/service/types.ts
@@ -1,21 +1,21 @@
-import { TokenManager } from '@backstage/backend-common';
-import { CatalogClient } from '@backstage/catalog-client';
+import {
+  PluginDatabaseManager,
+  PluginEndpointDiscovery,
+  TokenManager,
+} from '@backstage/backend-common';
+import { LoggerService } from '@backstage/backend-plugin-api';
 import { Config } from '@backstage/config';
 import { IdentityApi } from '@backstage/plugin-auth-node';
 import { PermissionEvaluator } from '@backstage/plugin-permission-common';
 
-import { Logger } from 'winston';
-
 export interface RouterOptions {
-  logger: Logger;
-  dbConfig: Config;
-  catalogClient: CatalogClient;
+  logger: LoggerService;
   identity: IdentityApi;
   permissions: PermissionEvaluator;
   tokenManager: TokenManager;
-
-  // Workaround - see auth.ts
-  externalCallerSecret?: string;
+  database: PluginDatabaseManager;
+  discovery: PluginEndpointDiscovery;
+  config: Config;
 }
 
 export type NotificationsFilterRequest = {
@@ -47,7 +47,7 @@ export type NotificationsFilterRequest = {
  */
 export type NotificationsSortingRequest = {
   orderBy?: string;
-  OrderByDirec?: string;
+  orderByDirec?: string;
 };
 
 export const NotificationsOrderByFields: string[] = [