+++ plugins/notifications/src/components/PersonalNotifications/PersonalNotifications.tsx
@@ -3,5 +3,5 @@ import React from 'react';
 import { NotificationsTable } from '../NotificationsTable';
 
 export const PersonalNotifications = () => (
-  <NotificationsTable messageScope="user" />
+  <NotificationsTable messageScope="user" title="Personal notifications" />
 );