+++ plugins/rbac-backend/src/service/role-manager.test.ts
@@ -41,12 +41,6 @@ describe('BackstageRoleManager', () => {
       ).toThrow('Method "syncedHasLink" not implemented.');
     });
 
-    it('should throw an error for getRoles', async () => {
-      await expect(roleManager.getRoles('name')).rejects.toThrow(
-        'Method "getRoles" not implemented.',
-      );
-    });
-
     it('should throw an error for getUsers', async () => {
       await expect(roleManager.getUsers('name')).rejects.toThrow(
         'Method "getUsers" not implemented.',
@@ -1121,6 +1115,75 @@ describe('BackstageRoleManager', () => {
       expect(result).toBeTruthy();
     });
   });
+
+  describe('getRoles returns roles per user', () => {
+    it('should returns role per user', async () => {
+      roleManager.addLink('user:default/test', 'role:default/rbac_admin');
+      roleManager.addLink('user:default/test-two', 'role:default/rbac_admin');
+      roleManager.addLink(
+        'user:default/test-three',
+        'role:default/rbac_admin_test',
+      );
+
+      catalogApiMock.getEntities.mockImplementation((arg: any) => {
+        const hasMember = arg.filter['relations.hasMember'];
+
+        if (hasMember && hasMember[0] === 'user:default/test') {
+          return { items: [] };
+        }
+        if (hasMember && hasMember[0] === 'user:default/test-two') {
+          return { items: [] };
+        }
+        if (hasMember && hasMember[0] === 'user:default/test-three') {
+          return { items: [] };
+        }
+        return { items: [] };
+      });
+
+      let roles = await roleManager.getRoles('user:default/test');
+      expect(roles.length).toBe(1);
+      expect(roles[0]).toEqual('role:default/rbac_admin');
+
+      roles = await roleManager.getRoles('user:default/test-two');
+      expect(roles.length).toBe(1);
+      expect(roles[0]).toEqual('role:default/rbac_admin');
+
+      roles = await roleManager.getRoles('user:default/test-three');
+      expect(roles.length).toBe(1);
+      expect(roles[0]).toEqual('role:default/rbac_admin_test');
+    });
+
+    it('getRoles returns role for user inherited from group', async () => {
+      const teamAGroup = createGroupEntity('team-a', undefined, [
+        'user:default/test',
+      ]);
+
+      roleManager.addLink('group:default/team-a', 'role:default/rbac_admin');
+
+      catalogApiMock.getEntities.mockImplementation((arg: any) => {
+        const hasMember = arg.filter['relations.hasMember'];
+
+        if (hasMember && hasMember[0] === 'user:default/test') {
+          return { items: [teamAGroup] };
+        }
+
+        return { items: [] };
+      });
+
+      let roles = await roleManager.getRoles('user:default/test');
+      expect(roles.length).toBe(1);
+      expect(roles[0]).toEqual('role:default/rbac_admin');
+
+      // should return empty array for group
+      roles = await roleManager.getRoles('group:default/team-a');
+      expect(roles.length).toBe(0);
+
+      // should return empty array for role
+      roles = await roleManager.getRoles('role:default/rbac_admin');
+      expect(roles.length).toBe(0);
+    });
+  });
+
   function createGroupEntity(
     name: string,
     parent?: string,