+++ plugins/notifications/CHANGELOG.md
@@ -1,3 +1,24 @@
+## @janus-idp/plugin-notifications [1.1.1](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/plugin-notifications@1.1.0...@janus-idp/plugin-notifications@1.1.1) (2024-01-24)
+
+
+### Bug Fixes
+
+* make the notifications icon active in dynamic plugin ([#1111](https://github.com/janus-idp/backstage-plugins/issues/1111)) ([bc98491](https://github.com/janus-idp/backstage-plugins/commit/bc984912f96adb8bc7575f154877dff7fcdbb6e8))
+
+## @janus-idp/plugin-notifications [1.1.0](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/plugin-notifications@1.0.3...@janus-idp/plugin-notifications@1.1.0) (2024-01-24)
+
+
+### Features
+
+* **notifications:** turn plugins into dynamic ones ([#1104](https://github.com/janus-idp/backstage-plugins/issues/1104)) ([5146977](https://github.com/janus-idp/backstage-plugins/commit/51469770c254a90790f466c3480add5caa41ca11))
+
+## @janus-idp/plugin-notifications [1.0.3](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/plugin-notifications@1.0.2...@janus-idp/plugin-notifications@1.0.3) (2024-01-24)
+
+
+### Bug Fixes
+
+* **notifications:** redesign the Notifications page ([#1095](https://github.com/janus-idp/backstage-plugins/issues/1095)) ([7efb107](https://github.com/janus-idp/backstage-plugins/commit/7efb107b6142ad966dd8cf836c8480e2316575b1))
+
 ## @janus-idp/plugin-notifications [1.0.2](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/plugin-notifications@1.0.1...@janus-idp/plugin-notifications@1.0.2) (2024-01-21)